change: truncate values of long string fields to database field length
in the email_log table, some fields (addresses and subject) have type
'string', which means laravel creates them with a default length of 255
characters.

however, in case of many e-mails - e.g. in cc - the total length of a
header can exceed this character limit. theoretically, a single e-mail
address could be [up to 320 characters
long](https://stackoverflow.com/a/49645137/3096626)

before this patch, the length limit was not checked and e-mails would
not be logged. error message:

```
[previous exception] [object] (PDOException(code: 22001): SQLSTATE[22001]: String data, right truncated: 1406 Data too long for column 'bcc' at row 1 at /var/www/vendor/doctrine/dbal/lib/Doctrine/DBAL/Driver/PDOStatement.php:115)
[stacktrace]
```

this patch truncates fields at 255 characters so the log entry can be
created. the untruncated contents are still available in the body
column.